% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{CalcNetworkFeatures}
\alias{CalcNetworkFeatures}
\title{Calculate Network and Node Level Statistics}
\usage{
CalcNetworkFeatures(edges, nodes)
}
\arguments{
\item{edges}{a data.table of edges with two columns, SOURCE, and TARGET}

\item{nodes}{a data.table of nodes with column node}
}
\value{
A two element list: \describe{
\item{networkMeasures}{a list of network level measures.}
\item{nodeMeasures}{a data.table with a node column andone column per node level statistic.}
}
}
\description{
Given an ExtractNetwork object, extract network features on a node and network level
}
