% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_report.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{Creates a}
\usage{
CreatePackageReport(packageName, packageReporters = DefaultReporters())
}
\arguments{
\item{pkgPath}{path to the package you want to examine}
}
\description{
Obtain Ratio of Coverage For Each Function Within A Package
}
\author{
B. Burns
}
