% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageVignette.R
\name{CreatePackageVignette}
\alias{CreatePackageVignette}
\title{pkgnet Report as Vignette}
\usage{
CreatePackageVignette(pkg_name, pkg_path = NULL,
  pkg_reporters = list(DependencyReporter$new(), FunctionReporter$new()),
  vignette_path = file.path("vignettes", "pkgnet_report.Rmd"))
}
\arguments{
\item{pkg_name}{(string) name of a package}

\item{pkg_path}{(string) The path to the package repository. If given, coverage
will be calculated for each function. \code{pkg_path} can be an
absolute or relative path.}

\item{pkg_reporters}{(list) a list of initialized package reporters}

\item{vignette_path}{(string) The path and filename of the output vignette
file. The default assumes your working directory is the package root.}
}
\description{
Create pkgnet package report as an Rmarkdown vignette. This
   vignette is able to be rendered with the
   \code{\link[knitr:vignette_engines]{knitr::rmarkdown}} vignette engine
   into HTML vignettes upon package building. It is also compatible with
   \code{\link[pkgdown:build_articles]{pkgdown}} sites.
}
