% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet.R
\name{doc_shared}
\alias{doc_shared}
\title{NULL Object For Common Documentation}
\arguments{
\item{pkg_name}{(string) name of a package}

\item{pkg_path}{(string) The path to the package repository. If given, coverage
will be calculated for each function. \code{pkg_path} can be an
absolute or relative path.}
}
\description{
This is a NULL object with documentation so that later functions can call
inheritParams
}
