% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_structure.R
\name{ExtractFunctionNetwork}
\alias{ExtractFunctionNetwork}
\title{Extract the Network Relationship Between Functions in A Package}
\usage{
ExtractFunctionNetwork(pkgName, pkgPath = NULL)
}
\arguments{
\item{pkgName}{(character) The name of the package from which to extract the network structure.}

\item{pkgPath}{(character) String, optional, with a path to the package's source code.
You need to provide this if you want to add test coverage
to the function nodes.}
}
\value{
A list with three (3) elements: \itemize{
\item{edges} {A data.table of dependencies between functions.}
}
}
\description{
This function maps the relationships between
             functions in a package. Optionally, a subset of functions
             can be mapped.
}
\examples{
\dontrun{
nw <- ExtractNetwork("data.table")
}
}
\seealso{
mvbutils::foodweb
}
\author{
B. Burns
}
