% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{GetCoverageByFunction}
\alias{GetCoverageByFunction}
\title{Obtain Ratio of Coverage For Each Function Within A Package}
\usage{
GetCoverageByFunction(pkgPath)
}
\arguments{
\item{pkgPath}{path to the package you want to examine}
}
\description{
Obtain Ratio of Coverage For Each Function Within A Package
}
\author{
B. Burns
}
