% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{MakeGraphObject}
\alias{MakeGraphObject}
\title{Make Graph Object Including Isolated Nodes}
\usage{
MakeGraphObject(edges, nodes)
}
\arguments{
\item{edges}{a data.table of edges with two columns, SOURCE, and TARGET}

\item{nodes}{a data.table of nodes with column node}
}
\value{
an igraph object
}
\description{
Given a pkgGraphect created by \code{\link{ExtractNetwork}},
             use \code{igraph} to create a formal graph object
}
