% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_structure.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Plot Network Structure}
\usage{
PlotNetwork(edges, nodes, colorFieldName = NULL)
}
\arguments{
\item{edges}{a data.table with SOURCE and TARGET nodes}

\item{nodes}{a data.table with node and augmenting metadata}

\item{colorFieldName}{(character) The name of column to use to color the nodes. This can
be any column in the "nodes" table inside the object produced by
\code{ExtractFunctionNetwork}. Default is outDegree. If you estimated
test coverage when creating the network representation,
try setting this field to "test_coverage"!}
}
\value{
A plotly object (local, not on plot.ly)
}
\description{
Plot the network structure of an R package.
}
\examples{
\dontrun{
nw <- ExtractFunctionNetwork("lubridate")
nw$nodes[,test := runif(.N)]
PlotNetwork(nw, colorFieldName = "test")
}
}
