% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{RenderPackageReport}
\alias{RenderPackageReport}
\title{Package Report Renderer}
\usage{
RenderPackageReport(reportPath, packageReporters)
}
\arguments{
\item{reportPath}{a file.path to where the report should be rendered}

\item{packageReporters}{a list of package reporters that have already been initialized and have calculated}
}
\value{
Nothing
}
\description{
Renders an html report based on the initialized reporters provided
}
\author{
P. Boueri
}
