% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractGraphReporter.R
\docType{data}
\name{AbstractGraphReporter}
\alias{AbstractGraphReporter}
\title{Abstract Graph Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractGraphReporter
}
\description{
Defines the Abstract Class for all PackageGraphReporters defined in pkgnet.
             The class is not meant to be instantiated, but inherited from and its methods
             overloaded such that each Metric implements certain functionality.
}
\section{Public Methods}{

\describe{
   \itemize{
       \item{\code{calculate_network_metrics}{Uses \code{CalcNetworkFeatures} to 
         calculate nodes and edges to create summary metrics of the overall 
         graph structure, including centrality measures etc.}}
       \item{\code{get_summary_view}{Returns the output of calculateNetworkMetrics}}
       \item{\code{set_graph_layout}{Accepts a layoutType from either "tree" or "circle"
         and augments the nodes private field with level/horizontal coordinates 
         for a prettified layout}}
       \item{\code{plot_network}{Calls \code{PlotNetwork} and returns a plotly object }}
   }
}
}

\section{Public Members}{

\describe{
 \itemize{
   \item{\code{edges}{A data.table from SOURCE to TARGET nodes describing the connections}}
   \item{\code{nodes}{A data.table with node as an identifier, and augementing information about each node}}
   \item{\code{pkgGraph}{An igraph object describing the package graph}}
   \item{\code{networkMeasures}{A list of network measures calculated by \code{CalcNetworkFeatures}}}
  }
}
}

\seealso{
Other AbstractReporters: \code{\link{AbstractPackageReporter}}
}
\keyword{datasets}
