% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractGraphReporter.R
\docType{data}
\name{AbstractGraphReporter}
\alias{AbstractGraphReporter}
\title{Abstract Graph Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractGraphReporter
}
\description{
Defines the Abstract Class for all PackageGraphReporters defined in pkgnet.
             The class is not meant to be instantiated, but inherited from and its methods
             overloaded such that each Metric implements certain functionality.
}
\section{Public Methods}{

\describe{
   \itemize{
        \item{\code{calculate_network_measures()}}{
            \itemize{
                \item{extract network features on a node and network level}
                \item{\bold{Returns:} A list containing:}{
                    \itemize{
                    \item{\bold{\code{networkMeasures}}: a list of network measures.}
                    \item{\bold{\code{nodeMeasures}}: the nodes data.table with additional columns of node measures.}
                   }
                }
            }
       }
       \item{\code{set_graph_layout}{Sets the layout of the graph plotted by
           \code{plot_network}. This method augments the nodes data.table
           with level/horizontal coordinates. The following layout types are
           supported:}}{\itemize{
               \item{`tree` (default)}
               \item{`circle`}
           }}
        \item{\code{plot_network}}{
            \itemize{
                \item{Creates a network visualization of extracted package graph.}
                \item{\bold{Args:}}{
                    \itemize{
                        \item{\bold{\code{...}}: ...}
                    }
                }
                \item{\bold{Returns:}}{
                    \itemize{
                        \item{A `visNetwork` object}
                   }
                }
            }
       }
   }
}
}

\section{Public Members}{

\describe{
 \itemize{
   \item{\code{edges}{A data.table from SOURCE to TARGET nodes describing the connections}}
   \item{\code{nodes}{A data.table with node as an identifier, and augmenting information about each node}}
   \item{\code{pkgGraph}{An igraph object describing the package graph}}
   \item{\code{networkMeasures}{A list of network measures calculated by \code{calculate_network_features}}}
   \item{\code{graphLayoutType}{Character string indicating currently active graph layout}}
   \item{\code{graphViz}{\code{visNetwork} object of package graph}}
  }
}
}

\seealso{
Other AbstractReporters: \code{\link{AbstractPackageReporter}}
}
\keyword{datasets}
