% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractGraphReporter.R
\docType{data}
\name{AbstractGraphReporter}
\alias{AbstractGraphReporter}
\title{Abstract Graph Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractGraphReporter
}
\description{
Defines the Abstract Class for all PackageGraphReporters defined in pkgnet.
             The class is not meant to be instantiated, but inherited from and its methods
             overloaded such that each Metric implements certain functionality.
}
\section{Public Methods}{

\describe{
   \itemize{
        \item{\code{calculate_network_measures()}}{
            \itemize{
                \item{extract network features on a node and network level}
                \item{\bold{Returns:}}{
                    \itemize{
                    \item{\bold{\code{packageName}}: String with the name of the package}
                    \item{\bold{\code{packagePath}}: Optional path to the source code. 
                        To be used for test coverage, if provided.}
                   }
                }
            }
       }
       \item{\code{get_summary_view}{Returns the output of calculateNetworkMetrics}}
       \item{\code{set_graph_layout}{Accepts a layoutType from either "tree" or "circle"
         and augments the nodes private field with level/horizontal coordinates 
         for a prettified layout}}

        \item{\code{plot_network(colorFieldName=NULL)}}{
            \itemize{
                \item{Creates a network visualization from tables of edges and nodes}
                \item{\bold{Args:}}{
                    \itemize{
                        \item{\bold{\code{packageName}}: The name of column to use to 
                            color the nodes. This can be any column in the "nodes" table
                            inside the object produced by \code{extract_network}.
                            Default is outDegree. If you estimated test coverage 
                            when creating the network representation, try setting 
                            this field to "test_coverage"}
                    }
                }
                \item{\bold{Returns:}}{
                    \itemize{
                        \item{A plotly object (local, not on plot.ly)}
                   }
                }
            }
       }
   }
}
}

\section{Public Members}{

\describe{
 \itemize{
   \item{\code{edges}{A data.table from SOURCE to TARGET nodes describing the connections}}
   \item{\code{nodes}{A data.table with node as an identifier, and augmenting information about each node}}
   \item{\code{pkgGraph}{An igraph object describing the package graph}}
   \item{\code{networkMeasures}{A list of network measures calculated by \code{CalcNetworkFeatures}}}
  }
}
}

\seealso{
Other AbstractReporters: \code{\link{AbstractPackageReporter}}
}
\keyword{datasets}
