% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractPackageReporter.R
\docType{data}
\name{AbstractPackageReporter}
\alias{AbstractPackageReporter}
\title{Abstract Package Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractPackageReporter
}
\description{
Defines the Abstract Class for all PackageReporters defined in pkgnet.
             The class is not meant to be instantiated, but inherited from and its methods
             overloaded such that each Metric implements certain functionality.
}
\section{Public Methods}{

\describe{
    \item{\code{set_package(pkg_name, pkg_path = NULL)}}{
        \itemize{
            \item{Set the package that all operations in the object are done for.}
            \item{\bold{Args:}}{
                \itemize{
                \item{\bold{\code{pkg_name}}: A string with the name of the package you are
                  analyzing.}
                \item{\bold{\code{pkg_path}}: Optional directory path to source
                  code of the package. It is used for calculating test coverage.
                  It can be an absolute or relative path.}
                 }
             }
         }
    }
    \item{\code{get_summary_view()}}{
        \itemize{
            \item{Returns a particular reporters summary report on the package for use in a high level view}
      }
    }
}
}

\seealso{
Other AbstractReporters: \code{\link{AbstractGraphReporter}}
}
\keyword{datasets}
