% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{Creates a}
\usage{
CreatePackageReport(packageName, packageReporters = DefaultReporters())
}
\arguments{
\item{packageName}{name of a package}

\item{packageReporters}{a list of package reporters}
}
\value{
A list of instantiated packageReporters the user can then interact with
}
\description{
Obtain Ratio of Coverage For Each Function Within A Package
}
\author{
B. Burns
}
