% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{Surface the internal and external dependencies of an R package.}
\usage{
CreatePackageReport(package_name, package_reporters = DefaultReporters(),
  package_path = NULL, report_path = file.path(getwd(), paste0(package_name,
  "_report.html")))
}
\arguments{
\item{package_name}{(string) name of a package}

\item{package_reporters}{(list) a list of package reporters}

\item{package_path}{(string) The path to the package repository.
If given, coverage will be calculated for each function.}

\item{report_path}{(string) The path and filename of the output report.  Default
report will be produced in working directory.}
}
\value{
A list of instantiated package_reporters fitted to \code{package_name}
}
\description{
Surface the internal and external dependencies of an R package.
}
\seealso{
GetPackageGraphs
}
\author{
B. Burns
}
