% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{Surface the internal and external dependencies of an R package.}
\usage{
CreatePackageReport(packageName, packageReporters = DefaultReporters(),
  packagePath = NULL)
}
\arguments{
\item{packageName}{name of a package}

\item{packageReporters}{a list of package reporters}

\item{packagePath}{(optional) the path to the package repository.
If given, coverage will be calculated for each function.}
}
\value{
A list of instantiated packageReporters the user can then interact with
}
\description{
Surface the internal and external dependencies of an R package.
}
\author{
B. Burns
}
