% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{Surface the internal and external dependencies of an R package.}
\usage{
CreatePackageReport(packageName, packageReporters = DefaultReporters(),
  packagePath = NULL, reportPath = file.path(getwd(), paste0(packageName,
  "_report.html")))
}
\arguments{
\item{packageName}{(string) name of a package}

\item{packageReporters}{(list) a list of package reporters}

\item{packagePath}{(string) The path to the package repository.
If given, coverage will be calculated for each function.}

\item{reportPath}{(string) The path and filename of the output report.  Default
report will be produced in working directory.}
}
\value{
A list of instantiated packageReporters fitted to \code{packageName}
}
\description{
Surface the internal and external dependencies of an R package.
}
\author{
B. Burns
}
