% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageReport.R
\name{CreatePackageReport}
\alias{CreatePackageReport}
\title{pkgnet Analysis Report for an R package}
\usage{
CreatePackageReport(pkg_name, pkg_reporters = DefaultReporters(),
  pkg_path = NULL, report_path = tempfile(pattern = pkg_name, fileext =
  ".html"))
}
\arguments{
\item{pkg_name}{(string) name of a package}

\item{pkg_reporters}{(list) a list of package reporters}

\item{pkg_path}{(string) The path to the package repository. If given, coverage
will be calculated for each function. \code{pkg_path} can be an
absolute or relative path.}

\item{report_path}{(string) The path and filename of the output report.  Default
report will be produced in the temporary directory.}
}
\value{
A list of instantiated pkg_reporters fitted to \code{pkg_name}
}
\description{
Create a standalone HTML report about a package and its networks.
}
