% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultReporters.R
\name{DefaultReporters}
\alias{DefaultReporters}
\title{Default Reporters}
\usage{
DefaultReporters()
}
\value{
list of instantiated reporter objects
}
\description{
Instantiates a list of default reporters to feed into
   \code{\link{CreatePackageReport}}.
}
