% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctionReporter.R
\docType{data}
\name{FunctionReporter}
\alias{FunctionReporter}
\title{Package Function Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FunctionReporter
}
\description{
This Reporter takes a package and uncovers the structure from
             its other functions, determining useful information such as which function is most
             central to the package. Combined with testing information it can be used as a powerful tool
             to plan testing efforts.
}
\section{Public Methods}{

\describe{
    \item{\code{set_package(pkg_name, pkg_path)}}{
        \itemize{
            \item{Set properties of this reporter. If pkg_name overrides a
                previously-set package name, any cached data will be removed.}
            \item{\bold{Args:}}{
                \itemize{
                \item{\bold{\code{pkg_name}}: String with the name of the package.}
                \item{\bold{\code{pkg_path}}: Optional directory path to source
                  code of the package. It is used for calculating test coverage.
                  It can be an absolute or relative path.}
               }
            }
        }
    }
}
}

\seealso{
Other PackageReporters: \code{\link{DependencyReporter}},
  \code{\link{SummaryReporter}}
}
\keyword{datasets}
