% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageInheritanceReporter.R
\docType{data}
\name{InheritanceReporter}
\alias{InheritanceReporter}
\title{Package Class Inheritance Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InheritanceReporter
}
\description{
Package Class Inheritance Reporter Class
}
\seealso{
Other PackageReporters: \code{\link{DependencyReporter}},
  \code{\link{FunctionReporter}},
  \code{\link{SummaryReporter}}
}
\concept{PackageReporters}
\keyword{datasets}
