% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDependencyReporter.R
\docType{data}
\name{PackageDependencyReporter}
\alias{PackageDependencyReporter}
\title{Package Dependency Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PackageDependencyReporter
}
\description{
This Reporter takes a package and uncovers the structure from
             its other package dependencies, determining which package it relies on is most central,
             allowing for a developer to determine how to vett its dependency tree
}
\section{Public Methods}{

\describe{
    \item{\code{set_package(packageName, ...)}}{
        \itemize{
            \item{Set the package that all operations in the object are done for.}
            \item{\bold{Args:}}{
                \itemize{
                \item{\bold{\code{packageName}}: a string with the name of the package you are
                  analyzing.}
                \item{\bold{\code{...}}: other arguments passed through to \code{ExtractDependencyNetwork}}
                 }
             }
         }
    }
}
}

\seealso{
Other PackageReporters: \code{\link{PackageFunctionReporter}}
}
\keyword{datasets}
