% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctionReporter.R
\docType{data}
\name{PackageFunctionReporter}
\alias{PackageFunctionReporter}
\title{Package Function Reporter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PackageFunctionReporter
}
\description{
This Reporter takes a package and uncovers the structure from
             its other functions, determining useful information such as which function is most 
             central to the package. Combined with testing information it can be used as a powerful tool
             to plan testing efforts.
}
\section{Dependency Methods}{

   \item{\code{setPackage}{Uses \link{ExtractFunctionNetwork} to create edges}}
   \item{\code{packageTestCoverage}{Uses \link{GetCoverageByFunction} to calculate node test coverage}}
}

\seealso{
Other PackageReporters: \code{\link{PackageDependencyReporter}}
}
\keyword{datasets}
