% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryReporter.R
\name{SummaryReporter}
\alias{SummaryReporter}
\title{Package Summary Reporter}
\description{
This reporter provides a high-level overview of a package via its package DESCRIPTION file.
}
\seealso{
Other Package Reporters: 
\code{\link{DependencyReporter}},
\code{\link{FunctionReporter}},
\code{\link{InheritanceReporter}}
}
\concept{Package Reporters}
\concept{Reporters}
\section{Super class}{
\code{\link[pkgnet:AbstractPackageReporter]{pkgnet::AbstractPackageReporter}} -> \code{SummaryReporter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{report_markdown_path}}{(character string) path to R Markdown template for this reporter. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SummaryReporter-get_summary_view}{\code{SummaryReporter$get_summary_view()}}
\item \href{#method-SummaryReporter-clone}{\code{SummaryReporter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pkgnet" data-topic="AbstractPackageReporter" data-id="set_package"><a href='../../pkgnet/html/AbstractPackageReporter.html#method-AbstractPackageReporter-set_package'><code>pkgnet::AbstractPackageReporter$set_package()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SummaryReporter-get_summary_view"></a>}}
\if{latex}{\out{\hypertarget{method-SummaryReporter-get_summary_view}{}}}
\subsection{Method \code{get_summary_view()}}{
Returns an htmlwidget object that summarizes the analysis of the reporter. 
Used when creating a \link[=CreatePackageReport]{package report}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SummaryReporter$get_summary_view()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SummaryReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SummaryReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SummaryReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
